Add ability to delete device action

+++ common/model/src/commonMain/kotlin/dev/bnorm/elevated/model/devices/DeviceActionArguments.kt
@@ -10,5 +10,5 @@ sealed class DeviceActionArguments
 @SerialName("PUMP_DISPENSE")
 class PumpDispenseArguments(
     val pump: Int,
-    val amount: Int, // milliliters
+    val amount: Double, // milliliters
 ) : DeviceActionArguments()
